Moved the regular expression file filter into a separate plugin.
