+++ docs/plugins/index.rst
@@ -63,6 +63,7 @@ Each plugin has its own set of options that can be defined in a section bearing 
    play
    plexupdate
    random
+   regexfilefilter
    replaygain
    rewrite
    scrub
@@ -151,6 +152,7 @@ Miscellaneous
 * :doc:`mbcollection`: Maintain your MusicBrainz collection list.
 * :doc:`missing`: List missing tracks.
 * :doc:`random`: Randomly choose albums and tracks from your library.
+* :doc:`regexfilefilter`: Automatically skip files during the import process based on regular expressions.
 * :doc:`spotify`: Create Spotify playlists from the Beets library.
 * :doc:`types`: Declare types for flexible attributes.
 * :doc:`web`: An experimental Web-based GUI for beets.
